tests: Improve zone max open handling
Make sure that handling of zone max open limit is conditional to the
existence of the support in the kernel, for null_blk settings and for
sysfs attribute file.

+++ tests/zonefs-tests-nullblk.sh
@@ -98,7 +98,10 @@ function create_zoned_nullb()
 		echo "$zone_capacity" > "$dev"/zone_capacity
 	fi
 	echo $1 > "$dev"/zone_nr_conv
-	echo "$zone_max_open" > "$dev"/zone_max_open
+
+	if [ -f "$dev"/zone_max_open ]; then
+		echo "$zone_max_open" > "$dev"/zone_max_open
+	fi
 
 	echo 1 > "$dev"/power
